fix(controller): fix the variables in the plan
